refactor(tests): one guard for the derived tables, not two
867473d added tests/test_derived_tables_match_their_inputs.py for the defect this
branch guards generally, so the same comparison ran twice by different routes:
that file for base_rate_precision.csv under pytest, the freshness guard for all
ten tables in the CI step where the results are checked. Two mechanisms for one
defect drift, and the one that goes stale is the one nobody is watching.

Absorbed rather than deleted. Its precision-column check is a different question
from anything the guard asks -- whether a derived column still agrees with the
columns it was computed from, which catches an edit or a changed formula with
every input sitting still -- so it moves across intact. Its docstring had the
consequence stated better than mine did: the stale rows put the headline "720
flags per true finding" out by more than a factor of three, which is now in this
file.

Its guard against a vacuous pass moves across too, generalised. The original
asserted that at least one row was compared; the version here asserts that at
least one TABLE was checked by either mechanism, since the way this guard fails
open is every table going unverifiable at once -- a renamed results directory, a
tool column that stops resolving -- with the suite staying green while nothing is
checked.
