+++ tests/test_table_freshness.py
@@ -5,7 +5,8 @@ checks the layer above: ``tables/*.csv`` and ``*.tex`` are arithmetic over those
 results, and nothing noticed when ``base_rate_precision.csv`` went on reporting
 a doi_only false-positive rate of 0.2788 after its source run was re-run at
 0.0417 — a table claiming a tool hands the user roughly seven times more false
-accusations than it does.
+accusations than it does, and putting its headline "720 flags per true finding"
+out by more than a factor of three.
 
 Two mechanisms, tested separately: recorded input hashes
 (:func:`record_table`), which works for a table of any shape, and a direct
@@ -299,4 +300,47 @@ def test_real_known_stale_table_register_only_shrinks():
     assert not obsolete, (
         f"listed in KNOWN_STALE_TABLES but no longer stale: {obsolete}. "
         "Remove them from the register."
+    )
+
+
+# --- Derived columns against their own inputs -------------------------------
+#
+# The checks above ask whether a table still agrees with the results it came
+# from. This one asks whether it agrees with ITSELF: a column computed from
+# other columns can be edited, or regenerated by a changed formula, without any
+# input moving. Absorbed from tests/test_derived_tables_match_their_inputs.py,
+# which covered this table alone.
+
+
+@pytest.mark.skipif(
+    not (_REAL_TABLES_DIR / "base_rate_precision.csv").is_file(),
+    reason="base-rate table not generated",
+)
+def test_precision_column_is_consistent_with_its_own_inputs():
+    """Bayes' rule, recomputed. Catches an edit to one column but not the others."""
+    with (_REAL_TABLES_DIR / "base_rate_precision.csv").open() as fh:
+        rows = list(csv.DictReader(fh))
+    bad: list[str] = []
+    for row in rows:
+        dr, fpr = float(row["detection_rate"]), float(row["false_positive_rate"])
+        p = float(row["prevalence"])
+        tp, fp = dr * p, fpr * (1 - p)
+        expected = tp / (tp + fp) if (tp + fp) > 0 else 0.0
+        if abs(float(row["precision"]) - expected) > 1e-3:
+            bad.append(f"{row['tool']} @ {p}: precision {row['precision']} != {expected:.4f}")
+    assert rows, "no rows were compared — the check would pass vacuously"
+    assert not bad, "precision column disagrees with its own DR/FPR:\n  " + "\n  ".join(bad)
+
+
+@pytest.mark.skipif(not _REAL_TABLES_DIR.is_dir(), reason="real tables dir not present")
+def test_the_real_run_actually_compares_something():
+    """A guard that checks nothing passes for the wrong reason.
+
+    Every table could go unverifiable — a renamed results dir, a tool column
+    that stops resolving — and the suite would stay green while nothing was
+    being checked at all.
+    """
+    res = crf.check_table_freshness(_REAL_TABLES_DIR, _REAL_RESULTS_DIR, repo_root=_REPO_ROOT)
+    assert any(r.value_checked or r.provenance_checked for r in res.reports), (
+        "no table was checked against its inputs: the guard is passing vacuously"
     )
